requirements.txt
faker
pytest
allure-pytest
selenium
